refactor(skills): promote skills to internal/skills
Move-only: the skills service implements agentdomain.SkillsService and owns
discovery, the catalog client, the installer and the embedded builtins. It is
its own context, not a service in a grab-bag.

The //go:embed builtins directive is directory-relative, so builtins/tmux/SKILL.md
travels with the package unchanged.
