+++ internal/skills/builtins_test.go
