+++ internal/skills/catalog.go
