+++ internal/skills/catalog_test.go
