+++ internal/skills/install.go
